Midterm/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.2
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.20.0
ruff==0.7.1
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
wheel==0.44.0
